Move stateflow to psql (#347)

+++ api/velour_api/backend/core/model.py
@@ -1,12 +1,84 @@
 import json
 
 from geoalchemy2.functions import ST_AsGeoJSON
-from sqlalchemy import select
+from sqlalchemy import and_, func, select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from velour_api import exceptions, schemas
 from velour_api.backend import models
+from velour_api.backend.core.annotation import create_skipped_annotations
+from velour_api.backend.core.dataset import fetch_dataset, get_dataset_status
+from velour_api.backend.core.evaluation import check_for_active_evaluations
+from velour_api.enums import ModelStatus, TableStatus
+
+
+def _fetch_disjoint_datums(
+    db: Session, dataset_name: str, model_name: str
+) -> list[models.Datum]:
+    """
+    Fetch all datums that the model has not provided predictions for.
+
+    Parameters
+    ----------
+    db : Session
+        The database session.
+    dataset_name : str
+        The name of the dataset.
+    model_name : str
+        The name of the model.
+
+    Returns
+    -------
+    list[models.Datum]
+        List of Datums.
+    """
+    dataset = fetch_dataset(db=db, name=dataset_name)
+    model = fetch_model(db=db, name=model_name)
+    disjoint_datums = (
+        select(models.Datum)
+        .join(
+            models.Annotation,
+            and_(
+                models.Annotation.datum_id == models.Datum.id,
+                models.Annotation.model_id == model.id,
+            ),
+            isouter=True,
+        )
+        .where(models.Datum.dataset_id == dataset.id)
+        .filter(models.Annotation.id.is_(None))
+        .subquery()
+    )
+    return db.query(disjoint_datums).all()
+
+
+def create_model(
+    db: Session,
+    model: schemas.Model,
+):
+    """
+    Creates a model.
+
+    Parameters
+    ----------
+    db : Session
+        The database Session to query against.
+    model : schemas.Model
+        The model to create.
+    """
+    try:
+        row = models.Model(
+            name=model.name,
+            meta=model.metadata,
+            geo=model.geospatial.wkt() if model.geospatial else None,
+            status=ModelStatus.READY,
+        )
+        db.add(row)
+        db.commit()
+        return row
+    except IntegrityError:
+        db.rollback()
+        raise exceptions.ModelAlreadyExistsError(model.name)
 
 
 def fetch_model(
@@ -35,34 +107,6 @@ def fetch_model(
     if model is None:
         raise exceptions.ModelDoesNotExistError(name)
     return model
-
-
-def create_model(
-    db: Session,
-    model: schemas.Model,
-):
-    """
-    Creates a model.
-
-    Parameters
-    ----------
-    db : Session
-        The database Session to query against.
-    model : schemas.Model
-        The model to create.
-    """
-    try:
-        row = models.Model(
-            name=model.name,
-            meta=model.metadata,
-            geo=model.geospatial.wkt() if model.geospatial else None,
-        )
-        db.add(row)
-        db.commit()
-        return row
-    except IntegrityError:
-        db.rollback()
-        raise exceptions.ModelAlreadyExistsError(model.name)
 
 
 def get_model(
@@ -109,7 +153,7 @@ def get_models(
     Parameters
     ----------
     db : Session
-        The database Session to query against.
+        The database session.
 
     Returns
     ----------
@@ -119,6 +163,141 @@ def get_models(
     return [
         get_model(db, name) for name in db.scalars(select(models.Model.name))
     ]
+
+
+def get_model_status(
+    db: Session,
+    dataset_name: str,
+    model_name: str,
+) -> TableStatus:
+    """
+    Get status of model.
+
+    Parameters
+    ----------
+    db : Session
+        The database session.
+    name : str
+        The name of the model.
+
+    Returns
+    -------
+    TableStatus
+        The status of the model.
+    """
+    dataset = fetch_dataset(db, dataset_name)
+    model = fetch_model(db, model_name)
+
+    # format statuses
+    dataset_status = TableStatus(dataset.status)
+    model_status = ModelStatus(model.status)
+
+    # check if deleting
+    if model_status == ModelStatus.DELETING:
+        return TableStatus.DELETING
+
+    # check dataset status
+    if dataset_status == TableStatus.DELETING:
+        raise exceptions.DatasetDoesNotExistError(dataset.name)
+    elif dataset_status == TableStatus.CREATING:
+        return TableStatus.CREATING
+
+    # query the number of datums that do not have any prediction annotations
+    query_num_disjoint_datums = (
+        select(func.count())
+        .select_from(models.Datum)
+        .join(
+            models.Annotation,
+            and_(
+                models.Annotation.datum_id == models.Datum.id,
+                models.Annotation.model_id == model.id,
+            ),
+            isouter=True,
+        )
+        .where(models.Datum.dataset_id == dataset.id)
+        .filter(models.Annotation.id.is_(None))
+    )
+
+    # finalization is determined by the existence of at least one annotation per datum.
+    if db.scalar(query_num_disjoint_datums) != 0:
+        return TableStatus.CREATING
+    else:
+        return TableStatus.FINALIZED
+
+
+def set_model_status(
+    db: Session,
+    dataset_name: str,
+    model_name: str,
+    status: TableStatus,
+):
+    """
+    Sets the status of a model.
+
+    Parameters
+    ----------
+    db : Session
+        The database session.
+    dataset_name : str
+        The name of the dataset.
+    model_name : str
+        The name of the model.
+    status : TableStatus
+        The desired dataset state.
+
+    Raises
+    ------
+    exceptions.DatasetDoesNotExistError
+        If the dataset does not exist or is being deleted.
+    exceptions.ModelStateError
+        If an illegal transition is requested.
+    exceptions.EvaluationRunningError
+        If the requested state is DELETING while an evaluation is running.
+    """
+    dataset_status = get_dataset_status(db, dataset_name)
+    if dataset_status == TableStatus.DELETING:
+        raise exceptions.DatasetDoesNotExistError(dataset_name)
+
+    model_status = get_model_status(db, dataset_name, model_name)
+    if status == model_status:
+        return
+
+    model = fetch_model(db, model_name)
+
+    # check if transition is valid
+    if status not in model_status.next():
+        raise exceptions.ModelStateError(model_name, model_status, status)
+
+    # verify model-dataset parity
+    if (
+        model_status == TableStatus.CREATING
+        and status == TableStatus.FINALIZED
+    ):
+        if dataset_status != TableStatus.FINALIZED:
+            raise exceptions.DatasetNotFinalizedError(dataset_name)
+        # edge case - check that there exists at least one prediction per datum
+        create_skipped_annotations(
+            db=db,
+            datums=_fetch_disjoint_datums(db, dataset_name, model_name),
+            model=model,
+        )
+
+    elif status == TableStatus.DELETING:
+        if check_for_active_evaluations(db=db, model_name=model_name):
+            raise exceptions.EvaluationRunningError(
+                dataset_name=dataset_name, model_name=model_name
+            )
+
+    try:
+        model.status = (
+            ModelStatus.READY
+            if status != TableStatus.DELETING
+            else ModelStatus.DELETING
+        )
+        db.commit()
+    except Exception as e:
+        db.rollback()
+        raise e
 
 
 def delete_model(
@@ -136,6 +315,16 @@ def delete_model(
         The name of the model.
     """
     model = fetch_model(db, name=name)
+    if check_for_active_evaluations(db=db, model_name=name):
+        raise exceptions.EvaluationRunningError(name)
+
+    try:
+        model.status = ModelStatus.DELETING
+        db.commit()
+    except Exception as e:
+        db.rollback()
+        raise e
+
     try:
         db.delete(model)
         db.commit()
